[IMP] migrated legacy test file in pytest
